server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["MrTj458"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
google-api-python-client = "^2.22.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
httpx = "^0.19.0"
aioredis = "^2.0.0"
asyncpg = "^0.24.0"
tortoise-orm = "^0.17.7"
aerich = "^0.5.8"

[tool.poetry.dev-dependencies]
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
